core/requirements.txt
sqlalchemy
asyncpg
python-dotenv
pyjwt
alembic
pytz
werkzeug
uvicorn
fastapi-restful
httpx
ujson
pydantic_settings
fastapi_swagger_dark
psycopg2-binary
typing_inspect
email-validator
pillow
python-multipart
pandas
xlsxwriter
fastapi-restful==0.6.0
elasticsearch==8.11.0
python-keycloak==0.35.0
python-jose[cryptography]==3.3.0

